+++ src/config_stuff/colors.rs
@@ -30,6 +30,9 @@ pub struct ColorsConfig {
 }
 
 pub struct Colors {
+    // If colors are actually need to be displayed
+    pub enabled: bool,
+
     pub name: (u8, u8, u8),
     pub underline: Option<(u8, u8, u8)>,
     pub id: (u8, u8, u8),
@@ -48,6 +51,8 @@ impl Colors {
     pub fn from_config(config: &Config) -> Self {
         let c = &config.colors;
         Colors {
+            enabled: true,
+
             name: c.name.map(Color::to_rgb).unwrap_or((203, 166, 247)),
             underline: c.underline.map(Color::to_rgb),
             id: c.id.map(Color::to_rgb).unwrap_or((137, 220, 235)),
@@ -60,5 +65,10 @@ impl Colors {
             twitter: c.twitter.map(Color::to_rgb).unwrap_or((203, 166, 247)),
             blog: c.blog.map(Color::to_rgb).unwrap_or((203, 166, 247)),
         }
+    }
+
+    pub fn no_color(mut self) -> Self {
+        self.enabled = false;
+        self
     }
 }