+++ src/config_stuff/config.rs
@@ -19,7 +19,7 @@ pub struct Config {
 pub fn load_config() -> Result<Config> {
     let path = match dirs::home_dir() {
         Some(home) => home.join(".config").join("ghfetch").join("config.toml"),
-        None => anyhow::bail!("could not find home directory"),
+        None => anyhow::bail!("Could not find home directory"),
     };
 
     if !path.exists() {
@@ -28,10 +28,10 @@ pub fn load_config() -> Result<Config> {
 
     // Show Error if the config fails to read or parse
     let contents = std::fs::read_to_string(&path)
-        .with_context(|| format!("failed to read {}", path.display()))?;
+        .with_context(|| format!("Failed to read {}", path.display()))?;
 
     let config = toml::from_str::<Config>(&contents)
-        .with_context(|| format!("failed to parse {}", path.display()))?;
+        .with_context(|| format!("Failed to parse {}", path.display()))?;
 
     Ok(config)
 }