+++ CHANGELOG.md
@@ -4,9 +4,18 @@ All notable changes to this project are documented here.
 
 ## [Unreleased]
 
+## Fixes
+
+- Made "break" field be removed when it is the last line in the fields to maintain consistency of one line gap.
+
+## Changed
+
+- Switched from a match to using clap for cli management.
+
 ### Added
 
 - Automated release builds for macOS and Linux.
+- `--no-avatar` and `--no-color` flags.
 
 ## [1.0.1] - 2026-09-22